requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.9.0
pytz==2024.2
sqlparse==0.5.1
tzdata==2024.2
drf-yasg